@@ -1,6 +1,6 @@
 {
   "name": "react-bootstrap",
-  "version": "0.24.3",
+  "version": "0.24.5",
   "description": "Bootstrap 3 components build with React",
   "repository": {
     "type": "git",
@@ -53,17 +53,18 @@
     "child-process-promise": "^1.1.0",
     "codemirror": "^5.5.0",
     "colors": "^1.1.2",
-    "css-loader": "^0.15.6",
+    "css-loader": "^0.16.0",
     "es5-shim": "^4.1.10",
-    "eslint": "1.0.0",
-    "eslint-plugin-babel": "^1.2.0",
+    "eslint": "^1.1.0",
+    "eslint-config-airbnb": "0.0.7",
+    "eslint-plugin-babel": "^2.0.0",
     "eslint-plugin-lodash": "^0.1.3",
     "eslint-plugin-mocha": "^0.4.0",
     "eslint-plugin-react": "^3.1.0",
     "express": "^4.13.1",
     "extract-text-webpack-plugin": "^0.8.2",
     "file-loader": "^0.8.4",
-    "fs-extra": "^0.22.1",
+    "fs-extra": "^0.23.0",
     "fs-promise": "^0.3.1",
     "glob": "^5.0.14",
     "http-proxy": "^1.11.1",
@@ -73,7 +74,7 @@
     "karma": "~0.13.3",
     "karma-chrome-launcher": "~0.2.0",
     "karma-cli": "0.1.0",
-    "karma-coverage": "^0.4.2",
+    "karma-coverage": "^0.5.0",
     "karma-coveralls": "^1.1.2",
     "karma-firefox-launcher": "~0.1.6",
     "karma-mocha": "~0.2.0",
