@@ -1,28 +1,40 @@
 import React from 'react';
 import classNames from 'classnames';
-import BootstrapMixin from './BootstrapMixin';
-import styleMaps from './styleMaps';
 
 const Glyphicon = React.createClass({
-  mixins: [BootstrapMixin],
-
   propTypes: {
-    glyph: React.PropTypes.oneOf(styleMaps.GLYPHS).isRequired
+    /**
+     * bootstrap className
+     * @private
+     */
+    bsClass: React.PropTypes.string,
+    /**
+     * An icon name. See e.g. http://getbootstrap.com/components/#glyphicons
+     */
+    glyph: React.PropTypes.string.isRequired,
+    /**
+     * Adds 'form-control-feedback' class
+     * @private
+     */
+    formControlFeedback: React.PropTypes.bool
   },
 
   getDefaultProps() {
     return {
-      bsClass: 'glyphicon'
+      bsClass: 'glyphicon',
+      formControlFeedback: false
     };
   },
 
   render() {
-    let classes = this.getBsClassSet();
-
-    classes['glyphicon-' + this.props.glyph] = true;
+    let className = classNames(this.props.className, {
+      [this.props.bsClass]: true,
+      ['glyphicon-' + this.props.glyph]: true,
+      ['form-control-feedback']: this.props.formControlFeedback
+    });
 
     return (
-      <span {...this.props} className={classNames(this.props.className, classes)}>
+      <span {...this.props} className={className}>
         {this.props.children}
       </span>
     );
