@@ -11,7 +11,11 @@ const Affix = React.createClass({
   mixins: [AffixMixin],
 
   render() {
-    let holderStyle = {top: this.state.affixPositionTop};
+    let holderStyle = {
+      top: this.state.affixPositionTop,
+      // we don't want to expose the `style` property
+      ...this.props.style // eslint-disable-line react/prop-types
+    };
 
     return (
       <div {...this.props}
