@@ -2,8 +2,6 @@ import React from 'react';
 import TransitionEvents from './utils/TransitionEvents';
 import deprecationWarning from './utils/deprecationWarning';
 
-let warned = false;
-
 const CollapsibleMixin = {
 
   propTypes: {
@@ -25,10 +23,7 @@ const CollapsibleMixin = {
   },
 
   componentWillMount(){
-    if ( !warned ){
-      deprecationWarning('CollapsibleMixin', 'Collapse Component');
-      warned = true;
-    }
+    deprecationWarning('CollapsibleMixin', 'Collapse Component');
   },
 
   componentWillUpdate(nextProps, nextState){
