@@ -1,20 +1,16 @@
-
 import React from 'react';
-import classnames from 'classnames';
+import classNames from 'classnames';
 
 class ModalHeader extends React.Component {
-
   render() {
     return (
       <div
         {...this.props}
-        className={classnames(this.props.className, this.props.modalClassName)}
-      >
+        className={classNames(this.props.className, this.props.modalClassName)}>
         { this.props.closeButton &&
           <button
             className='close'
-            onClick={this.props.onHide}
-          >
+            onClick={this.props.onHide}>
             <span aria-hidden="true">
               &times;
             </span>
@@ -35,14 +31,17 @@ ModalHeader.propTypes = {
    * It is used for Assistive Technology when the label text is not visible on screen.
    */
   'aria-label': React.PropTypes.string,
+
   /**
    * A css class applied to the Component
    */
   modalClassName: React.PropTypes.string,
+
   /**
    * Specify whether the Component should contain a close button
    */
   closeButton: React.PropTypes.bool,
+
   /**
    * A Callback fired when the close button is clicked. If used directly inside a Modal component, the onHide will automatically
    * be propagated up to the parent Modal `onHide`.
