@@ -64,7 +64,7 @@ export Portal from './Portal';
 export Position from './Position';
 
 export Collapse from './Collapse';
-export Fade from './Collapse';
+export Fade from './Fade';
 
 export * as FormControls from './FormControls';
 export * as utils from './utils';
